Fix stack overflow in MvxFormsPagePresenter.BuildPageTree
Assume the following setup:

A `NavigationPage` presenting a modal `MasterDetailPage`, which in turn has a `Detail` set to a NavigationPage. The second NP contains the MDP in its `ModalStack`, for whatever reason.

Now the current implementation recurses through the navigation stack, finds a `MasterDetailPage`, recurses through the `Detail` stack, finds a `NavigationPage`, checks the modal stack, finds a `MasterDetailPage`, recurses through the `Detail` stack... and oops, stack overflow. So I added another check to see if the modal is a MDP and is already containing the current `rootPage` in which case we just continue recursing happily.

As this is just adding 3 lines and removing 2, I thought to make a PR like this, if you need anything more please tell me, I am new to contributing to MvvmCross...